\name{voronoi.game}
\alias{voronoi.game}
\title{
Play the Voronoi game
}
\description{
A two player game where each player attempts to control as much space in the unit square, where space is controlled by those Voronoi cells owned by a player's points.
}
\usage{
voronoi.game(rounds)
}
\arguments{
  \item{rounds}{A scalar specifying how many rounds are to be played.}
}
\author{
Christopher D. Barr
}
\seealso{
\code{\link{fd}}
}
\examples{
# voronoi.game(4)
}
\keyword{ Voronoi cells }
\keyword{ Voronoi game }
